Move constants to common file
