+++ src/actionPanel/native.tsx
@@ -29,12 +29,11 @@ import {
   RendererError,
 } from "@/actionPanel/actionPanelTypes";
 import { RendererPayload } from "@/runtime/runtimeTypes";
-import { renderPanels, hideForm, showForm } from "@/actionPanel/messenger/api";
-import { MAX_Z_INDEX } from "@/common";
+import { hideForm, renderPanels, showForm } from "@/actionPanel/messenger/api";
+import { MAX_Z_INDEX, PANEL_FRAME_ID } from "@/common";
 import pDefer from "p-defer";
 
 const SIDEBAR_WIDTH_PX = 400;
-export const PANEL_FRAME_ID = "pixiebrix-extension";
 const PANEL_CONTAINER_SELECTOR = "#" + PANEL_FRAME_ID;
 export const PANEL_HIDING_EVENT = "pixiebrix:hideActionPanel";
